After Comp 12/2/17
There are some problems though
The red autos work fine but the blue ones have a lot of problems. Overall there was not a time where the red one failed. The blue one worked once after some changes durring comp but is overall not working. I am planning on making some small changes in the lab where the conditions of the feild are not being changed and i can quickly change things. - Gavin